@@ -14,7 +14,7 @@ description: A new Flutter application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -22,9 +22,9 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  equatable: ^1.1.1
-  permission_handler: ^5.0.0+hotfix.6
-  provider: ^4.1.2
+  equatable: ^2.0.3
+  permission_handler: ^8.1.2
+  provider: ^5.0.0
   contextualactionbar:
     path: ../
 
