@@ -10,19 +10,19 @@ class ContextualActionWidget<T> extends StatelessWidget {
   final Widget child;
 
   /// The color for the child when selected
-  final Color selectedColor;
+  final Color? selectedColor;
 
   /// The widget to be displayed on top of the child widget when selected
-  final Widget selectedWidget;
+  final Widget? selectedWidget;
 
   /// The widget to be displayed on top of
   /// the child widget when not selected
-  final Widget unselectedWidget;
+  final Widget? unselectedWidget;
 
   const ContextualActionWidget(
-      {Key key,
-      @required this.data,
-      @required this.child,
+      {Key? key,
+      required this.data,
+      required this.child,
       this.selectedWidget,
       this.unselectedWidget,
       this.selectedColor})
@@ -50,11 +50,12 @@ class ContextualActionWidget<T> extends StatelessWidget {
                           ? selectedColor ?? Theme.of(context).splashColor
                           : Colors.transparent),
                   child: child),
-              if (isItemAdded && selectedWidget != null) selectedWidget,
+              if (isItemAdded && selectedWidget != null)
+                selectedWidget ?? Container(),
               if (!isItemAdded &&
                   itemController.actionModeEnable &&
                   unselectedWidget != null)
-                unselectedWidget
+                unselectedWidget ?? Container()
             ],
           ),
         ),
